fix: Resolve remaining type inference errors in Swipe/Glide typing
Fixed issues:
- SwipeGestureDetector: Explicit Char.toInt() for range checks ('a'.toInt()..'z'.toInt())
- WordPredictor: Explicit Char.toInt() for range checks
- CustomKeyboardIME: Explicit CharSequence type for getSurroundingText return value

These fixes resolve Kotlin type inference issues with Char ranges and CharSequence handling

+++ app/src/main/java/com/example/customkeyboard/swipe/SwipeGestureDetector.kt
@@ -109,7 +109,7 @@ class SwipeGestureDetector(
     }
 
     private fun isLetterKey(code: Int): Boolean {
-        return code in 'a'..'z' || code in 'A'..'Z'
+        return code in 'a'.toInt()..'z'.toInt() || code in 'A'.toInt()..'Z'.toInt()
     }
 
     data class KeyPosition(
